Improve parameter transformation
